Introduced a separation between standard plugin copies and custom script
