+++ function/hls/lambda.yaml
@@ -3,6 +3,7 @@
 install:
     - null 
 setup:
+    - "python ./runtime_env/hls/scripts/base.py %s"
     - "python ./runtime_env/hls/scripts/main.py %s"
 run:
     - null
