+++ function/input/lambda.yaml
@@ -3,6 +3,6 @@
 install:
     - null
 setup:
-    - "python runtime_env/rest/scripts/main.py %s"
+    - "python runtime_env/rest/scripts/base.py %s"
 run:
     - null
